PWA-2415: Refactored `Form/Builder` component.
